Refactor: Clear all aislop findings
The aislop scanner reported 66 findings (score 39/100). This clears
every one of them, taking the repository to 100/100, without changing
observable behaviour.

Workflow output (43 findings). The print() calls flagged as debug
leftovers were GitHub Actions job-log writes: workflow commands such
as ::group:: and ::notice:: plus the human-readable run transcript.
That output is part of the action's contract with the runner, so it
cannot move to the logger, which writes to stderr and can be silenced
by log level. A new workflow_output module names the protocol and owns
the stdout writes, so a bare print elsewhere in the package once again
means what the rule assumes it means.

Oversized modules (2 findings). cli_interface.py keeps the Typer
option surface; the increment pipeline it drives moves to
increment_command.py. VersionIncrementer splits along an existing
seam into VersionIncrementerBase (known-tag state and lookups),
PrereleaseIncrementer (pre-release strategy), and the concrete class
(public API plus major/minor/patch). The split is pure code motion:
every method stays on the instance, so attribute access and test
patch targets are unchanged.

Long function (1 finding). _execute_increment splits into helpers for
resolving the version source, reading settings, collecting tags, and
applying the increment, with the settings now a typed dataclass.

Trivial comments (17 findings). Hand-reviewed rather than auto-fixed:
pure restatements are gone, and the comments carrying ordering or
performance rationale are reworded to explain why instead of what.

Formatting (3 findings) is ruff format output.

Review feedback follows.

The workflow command emitters now escape the text they interpolate:
percent, CR and LF in a message, and additionally ':' and ',' in a
property value. Tag names and exception strings reach these helpers
from outside the process, so without escaping a newline would end the
command and let the rest of the string open a new one. echo() stays
verbatim, because escaping there would mangle the '::' markers of the
commands passed to it.

The exhausted-search fallback in _create_first_prerelease built its
prerelease from the caller's prerelease_type rather than the resolved
prerelease_id, so an unset type produced the version "1.2.3-None.1".
It now uses prerelease_id, matching the search above it. The bug
predates this refactor and was carried over by the code motion.

The docstring of the same function described a fallback that
increments the patch, but the fallback keeps the unincremented patch.
It now describes the forward search and the fallback accurately.

A duplicated comment in _increment_existing_prerelease collapses to a
single line explaining why numbering restarts under the next patch
instead of restating the conditional below it.

Tests cover both behavioural fixes: the new tests/test_workflow_output
module asserts the escaping of each emitter, and two cases in
test_incrementer.py exhaust the prerelease search to reach the
fallback with and without an explicit type.
